refactor(comments): move plugin executable to dev
